Crate trampoline

Trampoline CLI

Welcome to Trampoline CLI documentation!

Trampoline CLI is a command-line interface that provides a host of capabilities for running & configuring a local development environment, creating and managing projects, and testing smart contracts. Trampoline currently provides the following features:

Quickly generate new dapp projects

Launch local nodes, miners, and chain indexers

Generate custom schemas for on-chain data

Installation

Currently, Trampoline has only been tested on Linux environments.

Trampoline uses docker for running chain services like nodes and miners, so it is recommended to install docker before moving on.

To install trampoline crate run: (currently not supported)

cargo install trampoline

To install from github run:

cargo install --git https://github.com/Tempest-Protocol/trampoline --branch develop

Usage

Please note that Trampoline is in early stages and under active development. The API can change drastically.

USAGE:
    trampoline <SUBCOMMAND>
FLAGS:
    -h, --help       Prints help information
    -V, --version    Prints version information
SUBCOMMANDS:
    help       Prints this message or the help of the given subcommand(s)
    network    Manage local development chain
    new        Create a new Trampoline project
    schema     Manage custom on chain structures

Quick Start

After installation, use the commands below to create a new trampoline project and manage a local network.

trampoline new example

Click here to see the output.

{{#markdown}} ```bash PATH: generators/Cargo.toml PATH: schemas/src/lib.rs PATH: DockerIndexer PATH: schemas/Cargo.toml PATH: trampoline.toml PATH: generators/src/lib.rs PATH: src/lib.rs PATH: Cargo.toml PATH: Dockerfile.template PATH: trampoline-env.toml [+] Building 0.9s (5/5) FINISHED => [internal] load build definition from Dockerfile 0.0s => => transferring dockerfile: 84B 0.0s => [internal] load .dockerignore 0.0s => => transferring context: 2B 0.0s => [internal] load metadata for docker.io/nervos/ckb:v0.103.0 0.8s => CACHED [1/1] FROM docker.io/nervos/ckb:v0.103.0@sha256:a4ca46fee4f19ee383270def46044055dbcca1373eb2e1075eba821d090635cd 0.0s => exporting to image 0.0s => => exporting layers 0.0s => => writing image sha256:5069d16d6a48b181e4a8a9b898ef7cf701c1e16a63c490c0bb0ede12eb01165c 0.0s => => naming to docker.io/tempest/trampoline-env:latest 0.0s Use 'docker scan' to run Snyk tests against images to find vulnerabilities and learn how to fix them ``` {{/markdown}}

cd example

trampoline network init

Click here to see the output.

{{#markdown}} ```bash New Trampoline development network inititalized. Network name: example-network Network ID: 6ad5c643b79d582cc40694541cc9a426bc43c3e9b54265f5bd39668ffdf61494 Network Services: ["example-ckb", "example-ckb-indexer"] ``` {{/markdown}}

trampoline network launch

Click here to see the output.

{{#markdown}} ```bash Nodes should have started! Indexer should have started! ``` {{/markdown}}

Project Layout

trampoline.toml - Trampoline project configuration file

network.toml - Network services configuration file

generators - Transaction generator directory

schemas - Custom cell schema directory

scripts - Smart contracts directory

.trampoline - Directory for caching local chain & indexer data

Manage Local Network

trampoline network init - Initialize a new network

trampoline network launch - Launch the network

trampoline network miner - Start the miner

trampoline network indexer - Start an indexer

Manage Schemas

Schemas use the molecule encoding & serialization format.

trampoline schema new <schema_file> - Create a new schema

trampoline schema new byte_10_arr "array my_array [byte;10]" - Optionally, pass a schema definition (in Molecule)

trampoline schema build <schema_name> - Generate rust bindings to build and decode schema

Re-exports

pub use network::docker;

Modules

network

Types for managing a local network

opts

Definitions for command line interface

project

Types for managing trampoline projects

schema

Types for managing schemas

Structs

TEMPLATES

This creates the template files for a trampoline project.

Enums

TrampolineResourceType

Enumeration of a trampoline project and schema.

Statics

DAPP_FILES

Traits

TrampolineResource

Interface for TrampolineResource.

Functions

hex_decode

Decode hex

hex_encode

Encode hex

hex_string

Convert a hex to a string. Adapted from HexParser in ckb-cli/utils/arg_parser

parse_hex

Parse a hex string to a vector